crdbtest: move benchmarks to separate file
